Add non-subpixel antialiased font smoothing to all icons
Applies to #5

The fonts in the tabs weren't bold, they were using subpixel
antialiasing. This is the default and is really great for text, but not
so much for icon fonts. It turns out that there was something that was
causing the icons in the tree view to use just regular antialiasing,
but not in the tabs. This fixes that.

See this article for why regular antialiasing is better for the icon
fonts than subpixel-antialiasing:

https://github.com/blog/1135-the-making-of-octicons